DEXW-1982: move rate estimator to services
